function to print elements of list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int gloryCount;
+	size_t gloryCount;
 
 	gloryCount = 0;
